+++ packages/plylog/CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## Unreleased
 
 <!-- New PRs should document their changes here. -->
+* Add the ability to provide a new default transport.
+* [Breaking] Log level is now a property on `defaultConfig`, not on the module export directly.
+
 
 ## [0.3.0] - 2016-05-13